requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.4
pydantic-settings==2.7.1
httpx==0.28.1
python-multipart==0.0.20

pytesseract==0.3.13
Pillow==11.1.0

sentence-transformers==3.3.1
chromadb==0.5.23
torch==2.5.1

langsmith==0.2.10
langchain==0.3.14

openai==1.59.3

pytest==8.3.4
pytest-cov==6.0.0
